+++ Scripts/AnimatableProperty.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace Coffee.UIExtensions
 {