+++ Scripts/BakingCamera.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace Coffee.UIExtensions
+namespace Coffee.UIParticleExtensions
 {
     internal class BakingCamera : MonoBehaviour
     {