+++ Scripts/Editor/AnimatedPropertiesEditor.cs
@@ -1,4 +1,3 @@
-using System;
 using UnityEditor;
 using UnityEngine;
 using System.Collections.Generic;